refactor: components 디렉터리 내부 구조 변경 (2)
